refactor(demos): migrate wave 4 (the-form, values, errors) to the shared style system
Adds three registry pieces and migrates the readout-grid trio onto them:
- `layout` gains a `split` modifier for the common responsive two-column
  layout (stacked on narrow, side by side from 760px up).
- a `section` fragment for titled readout blocks (a heading over its content).
- a `deflist` fragment for `dl`/`dt`/`dd` leaf-read lists, with the `dd`
  rendered as the same light box as `pre`.

the-form, values and errors drop their per-demo grid, panel and readout CSS
for these fragments; their dark terminal `pre`/`dd` boxes become the canonical
light readout. the-form's submit stays accent via `button.primary` inside the
neutral `actions` row.

Also retrofits preprocess onto `layout split` + `section`, removing the
duplicate two-column and section CSS it had been carrying as residue.

Smoke green (63 passed, 4 skipped).
